AnyToken + AnyTag removed, some code refactored
