+++ CompileLib/Common/BufferedEnumerator.cs
@@ -7,6 +7,10 @@ using System.Diagnostics;
 
 namespace CompileLib.Common
 {
+    /// <summary>
+    /// Enumerator with restarting
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
     internal class BufferedEnumerator<T> where T : new()
     {
         private readonly LinkedList<T> buffer = new();
@@ -22,6 +26,10 @@ namespace CompileLib.Common
             current = start;
         }
 
+        /// <summary>
+        /// Move to next character
+        /// </summary>
+        /// <returns></returns>
         public bool MoveNext()
         {
             if (current is null)
@@ -48,6 +56,9 @@ namespace CompileLib.Common
             return true;
         }
 
+        /// <summary>
+        /// Get current character
+        /// </summary>
         public T Current
         {
             get
@@ -57,11 +68,19 @@ namespace CompileLib.Common
             }
         }
 
+        /// <summary>
+        /// Restart reading the character sequence
+        /// </summary>
         public void Restart()
         {
             current = start;
         }
 
+        /// <summary>
+        /// Remove the given number of character making impossible to restart reading from any of them
+        /// </summary>
+        /// <param name="count"></param>
+        /// <returns></returns>
         public T[] Extract(int count)
         {
             var result = new T[count];
