+++ CompileLib/ParserTools/IErrorHandler.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace CompileLib.ParserTools
 {
+    /// <summary>
+    /// Internal error handler interface
+    /// </summary>
     internal interface IErrorHandler
     {
         void Handle(object?[] prefix, ErrorHandlingDecider decider);
