+++ CompileLib/ParserTools/IProductionHandler.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace CompileLib.ParserTools
 {
+    /// <summary>
+    /// internal production handler interface
+    /// </summary>
     internal interface IProductionHandler
     {
         object? Handle(object?[] children);
