--- CompileLib/Parsing/AnyTagAttribute.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CompileLib.Parsing
-{
-    /// <summary>
-    /// [Only in ErrorProductions] Requires any tag
-    /// </summary>
-    [AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false)]
-    public class AnyTagAttribute : Attribute
-    {
-    }
-}
