+++ CompileLib/Parsing/ParsingException.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -12,6 +13,17 @@ namespace CompileLib.Parsing
     public class ParsingException : Exception
     {
         public ParsingException(string msg) : base(msg)
-        { }
+        {
+        }
+
+        public ParsingException(MethodInfo method, ParameterInfo parameter, string message)
+            : this($"[Method {method.DeclaringType.Name}::{method.Name}, parameter {parameter.Name}] {message}")
+        {
+        }
+
+        public ParsingException(MethodInfo method, string message)
+            : this($"[Method {method.DeclaringType.Name}::{method.Name}] {message}")
+        {
+        }
     }
 }
