+++ CompileLib/Parsing/SingleAttribute.cs
@@ -7,10 +7,11 @@ using System.Threading.Tasks;
 namespace CompileLib.Parsing
 {
     /// <summary>
-    /// Single child
+    /// Attribute to define single child
     /// </summary>
     [AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false)]
     public class SingleAttribute : Attribute
     {
+        public static readonly SingleAttribute Instance = new();
     }
 }
