+++ TestCompiler/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "TestCompiler": {
       "commandName": "Project",
-      "commandLineArgs": "testError.txt result.exe"
+      "commandLineArgs": "test2.txt result.exe"
     }
   }
 }
